fix: simplify the test
